include gzip compression and set as default

+++ autosubmit/log/utils.py
@@ -16,9 +16,13 @@
 # along with Autosubmit.  If not, see <http://www.gnu.org/licenses/>.
 
 import lzma
+import gzip
 import re
 from pathlib import Path
 from typing import Optional
+
+XZ_MAGIC = "FD 37 7A 58 5A 00"
+GZIP_MAGIC = "1F 8B"
 
 
 def compress_xz(
@@ -46,19 +50,53 @@ def compress_xz(
         with lzma.open(output_path, "wb", preset=final_preset) as output_file:
             output_file.writelines(input_file)
 
-    if not keep_input:
+    if not keep_input and input_path != output_path:
+        Path(input_path).unlink(missing_ok=True)
+
+
+def compress_gzip(
+    input_path: str,
+    output_path: str = None,
+    compression_level: int = 9,
+    keep_input: bool = True,
+):
+    """
+    Compress a file using Gzip compression.
+
+    :param input_path: Path to the input file.
+    :param output_path: Path to the output compressed file. If None, defaults to <input_path>.gz.
+    :param compression_level: Compression level (0-9). Defaults to 9.
+    :param keep_input: Whether to keep the original input file. Defaults to True.
+    """
+
+    if output_path is None:
+        output_path = f"{input_path}.gz"
+
+    with open(input_path, "rb") as input_file:
+        with gzip.open(
+            output_path, "wb", compresslevel=compression_level
+        ) as output_file:
+            output_file.writelines(input_file)
+
+    if not keep_input and input_path != output_path:
         Path(input_path).unlink(missing_ok=True)
 
 
 def is_xz_file(filepath: str):
     with open(filepath, "rb") as f:
         magic = f.read(6)
-    return magic == bytes.fromhex("FD 37 7A 58 5A 00")
+    return magic == bytes.fromhex(XZ_MAGIC)
+
+
+def is_gzip_file(filepath: str):
+    with open(filepath, "rb") as f:
+        magic = f.read(2)
+    return magic == bytes.fromhex(GZIP_MAGIC)
 
 
 def find_uncompressed_files(file_path: str, pattern: Optional[str] = None) -> list[str]:
     """
-    Return all files that are not compressed with xz in a directory and 
+    Return all files that are not compressed with xz in a directory and
     match the filename with the given regex pattern.
     """
 
@@ -67,9 +105,7 @@ def find_uncompressed_files(file_path: str, pattern: Optional[str] = None) -> li
 
     # Get all files in the directory sorted by modification time
     all_files = sorted(
-        Path(file_path).glob("*"),
-        key=lambda f: f.stat().st_mtime,
-        reverse=True
+        Path(file_path).glob("*"), key=lambda f: f.stat().st_mtime, reverse=True
     )
 
     result = []
@@ -79,7 +115,7 @@ def find_uncompressed_files(file_path: str, pattern: Optional[str] = None) -> li
             continue
 
         # Check if the file is not compressed
-        if not is_xz_file(str(filename)):
+        if not is_xz_file(str(filename)) and not is_gzip_file(str(filename)):
             result.append(str(filename))
 
     return result
